fixes bug with requesting cell at point which is out of cell range. And other
In lines 72 is the bug fixing. In line 39 we see that I added "0.1f"
parameter to getPointsFromShape() flag, so that we use only 10% of
points from the shape. (Because original shape of cat has like 300
points and it's not necessary. 30-70 points should be still enough.
Maybe in future need to set fixed number of points as parameter).