moves geoToPixel() inside of the Table class
because calculation of screen location of the geo-coordinates is now
happening when class Table is loaded, it is better to seal this method
inside of the Table class. Logically it belongs there. But to do that we
need to pass "mapScreenWidth" and "mapScreenHeight" variables to Table
class, as they are used to do geoToPixel() conversion. Alltogether this
reduces cohesion of the elements of the sketch and gives clearer code.
We still have mapGeoXXXX variables to get rid of.