fixes the ProcessingJS bug and tidies up some comments.
the problem was that when running in ProcessingJS instead of
getFloat(int, int) was called getFloat(String, int) and same thing
happened with getString(). So intead of getString(int, int) PJS would
call getString(String, int). This is why in PJS program would run
completely different pieces of code.
The solution is to renate getFloat(String, int) to
getFloatByName(String, int) and to rename getString(String, int) to
getStringByName(String,int);